Move client stuff into own folder
